Replaces audio with AudioPlayer in order to store delay effect on instance
